🏠 Build Command to Class with Standardized Warnings Part 1 (#544)
* refactor: base class for build command

* refactor: moved build command into class

* refactor: added warning

* refactor: moved over into build class

* refactor: switch over build_congif

* refactor: Switch over process_config_files

* refactor: remove old build_config
