+++ cognite_toolkit/_cdf_tk/user_warnings.py
@@ -92,6 +92,44 @@ class GeneralWarning(ToolkitWarning, ABC):
 
 
 @dataclass(frozen=True)
+class UnexpectedFileLocationWarning(ToolkitWarning):
+    severity: ClassVar[SeverityLevel] = SeverityLevel.LOW
+    filepath: str
+    alternative: str
+
+    def get_message(self) -> str:
+        message = f"{self.filepath!r} does not exist. Using {self.alternative!r} instead."
+        return SeverityFormat.get_rich_severity_format(self.severity, message)
+
+
+@dataclass(frozen=True)
+class LowSeverityWarning(GeneralWarning):
+    severity: ClassVar[SeverityLevel] = SeverityLevel.LOW
+    message_raw: str
+
+    def get_message(self) -> str:
+        return SeverityFormat.get_rich_severity_format(self.severity, self.message_raw)
+
+
+@dataclass(frozen=True)
+class MediumSeverityWarning(GeneralWarning):
+    severity: ClassVar[SeverityLevel] = SeverityLevel.MEDIUM
+    message_raw: str
+
+    def get_message(self) -> str:
+        return SeverityFormat.get_rich_severity_format(self.severity, self.message_raw)
+
+
+@dataclass(frozen=True)
+class HighSeverityWarning(GeneralWarning):
+    severity: ClassVar[SeverityLevel] = SeverityLevel.HIGH
+    message_raw: str
+
+    def get_message(self) -> str:
+        return SeverityFormat.get_rich_severity_format(self.severity, self.message_raw)
+
+
+@dataclass(frozen=True)
 class ToolkitDependenciesIncludedWarning(GeneralWarning):
     severity: ClassVar[SeverityLevel] = SeverityLevel.LOW
     message: ClassVar[str] = "Some resources were added due to dependencies."
