+++ tests/tests_unit/test_cli/test_behavior.py
@@ -9,9 +9,9 @@ from cognite.client.data_classes import Transformation, TransformationWrite
 from pytest import MonkeyPatch
 
 from cognite_toolkit._cdf import build, deploy, dump_datamodel_cmd, pull_transformation_cmd
+from cognite_toolkit._cdf_tk.build import BuildCommand
 from cognite_toolkit._cdf_tk.exceptions import ToolkitDuplicatedModuleError
 from cognite_toolkit._cdf_tk.load import TransformationLoader
-from cognite_toolkit._cdf_tk.templates import build_config
 from cognite_toolkit._cdf_tk.templates.data_classes import BuildConfigYAML, Environment, SystemYAML
 from cognite_toolkit._cdf_tk.utils import CDFToolConfig
 from tests.tests_unit.approval_client import ApprovalCogniteClient
@@ -67,7 +67,7 @@ def test_duplicated_modules(build_tmp_path: Path, typer_context: typer.Context) 
     config.environment.name = "dev"
     config.environment.selected_modules_and_packages = ["module1"]
     with pytest.raises(ToolkitDuplicatedModuleError) as err:
-        build_config(
+        BuildCommand().build_config(
             build_dir=build_tmp_path,
             source_dir=PROJECT_WITH_DUPLICATES,
             config=config,
